Domain conversion to objectify round 1
Former-commit-id: 116d2fcb1214a70b156c934a0996e5269d89c9f6

+++ gwtp-sample-carstore/src/main/java/com/arcbees/carsample/client/application/rating/RatingView.java
@@ -77,7 +77,7 @@ public class RatingView extends ViewWithUiHandlers<RatingUiHandlers> implements 
     private void initDataColumns() {
         Column<Rating, Number> idColumn = new Column<Rating, Number>(new NumberCell()) {
             @Override
-            public Integer getValue(Rating rating) {
+            public Long getValue(Rating rating) {
                 return rating.getId();
             }
         };
